Don't assume that the term slug matches the sanitized version of the term name.
If a term "foo" already exists before the plugin is installed, and then the plugin wants to create a term "foo" in its own taxonomy, the slug will be "foo-2". Querying for posts with slug "foo" will return nothing. Instead, we lookup the slug based on the name and taxonomy, so we can be sure to get the correct slug.