Improvements to the responsive component of MP6, props tollmanz.
---

Improve usage of Backbone.js for the menu toggle.

* Make the "Menu" label translatable.
* Use an underscore template for building the button. Underscores templates greatly improve maintainability and readability. Should another attribute need to be added to the template, setting 
this up will make that a very easy maintenance task. It also removes the templating from the rendering logic.
* The previous version created a menu view when the viewport width was narrow, then destroyed it when it was widened. This is unnecessarily complex and costly. This view should be thought of as 
a singleton in that a single menu is only ever needed for a page view. Instead of creating and destroying, it should only be created, then hidden and show. This is much friendly from a 
performance standpoint and removes the possibility of the dreaded zombie views, which the previous code went to great length to attempt to prevent.
* Completely decouple the view from the 'moby6' object. The beauty of Backbone views is that it promotes decoupled code. Using the 'moby6' object defeats this purpose and does not allow it to 
stand on its own Additionally, the previous code created a new 'div' prepended it to a DOM node, then created a new element in the view and prepended it to the prepended node. It is best to 
create the new node and preform a single write instead of multiple writes. All of this logic should also be within the view.
* For performance and readability reasons, make additional objects available to the IIFE.

'_self' is not used in 'init', so remove it.

Add more selector caching where necessary.

Make sure the function returns the same type all of the time.

The menu button load time is improved.

The menu load time is taking approximately 1100-1200ms to load. On an iPad, where this is needed, there is a noticeable flash of the menu loading in after the admin bar fades in. This patch 
completely eliminates that flash and improves the load time by roughly 500ms to 600-700ms.

There were three things affecting the load time:

* The script was loaded in the header
* 'moby6' was init'd on 'document ready'
* The 'matchMedia' function, which activates the menu, included a timeout that is appropriate for attaching it to the resize event; however, this was also applied to the initial load, which 
delayed the load 150ms

Each of these three items have been improved:

* The script is loaded in the footer
* Because the script is loaded in the footer, 'moby6' can be initialized without delay (~350ms gain)
* The 'matchMedia' function was refactored: * The function was renamed to 'toggleMenu' to more appropriately   classify what the function does * The interval was decoupled from the menu toggle 
functionality. This   allows us to use the 'toggleMenu' function for multiple purposes   and not incur the timeout penalty. * Underscore.js' 'debounce' function was used to setup a responsible   
throttling of the resize event. The previous iteration was   throttling; however, I changed this as it is a cleaner   implementation, which draws on a community vetted 'debounce' utility.