The Facebook Graph API started returning "stories", which are notifications rather than normal posts. For example, "John Doe likes a photo" is a "story".
The software had a bug where if the user's feed contained a story, that
item would output a blank entry. This caused confusion for users,
especially those with the count set to 1 (as this would mean that
randomly no posts at all would appear).

This bug is fixed with this commit. Now, we are always requesting as
many items as possible from the Graph API, and then filtering until we
find enough valid items to return to the user. This ensures that the
number of posts returned matches the number requested, whenever
possible.